# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Propolis Fulfillment API",
        "description": "Dokumentasi resmi layanan backend Propolis. Semua response menggunakan format {success, data, message}.",
        "contact": {
            "name": "Tim Propolis",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Root"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/admin/orders": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Daftar order",
                "operationId": "38dbfcc9c8ca1dc1ea794f1e7143d9c1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar order"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/manual": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Buat pesanan manual (offline/marketplace)",
                "operationId": "d2d4fff1a2412b4db4fb70c787aee4b4",
                "responses": {
                    "201": {
                        "description": "Pesanan manual berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Detail order untuk admin",
                "operationId": "d0ebbbb1b2a3aee324b8e64826b88b6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail order"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/verify-payment": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Verifikasi pembayaran dan alokasi batch",
                "operationId": "b5870a606a5c91804126c08682a883cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order dialokasikan"
                    },
                    "422": {
                        "description": "Status tidak valid"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/product-codes": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Input atau perbarui kode produk untuk setiap item order",
                "operationId": "3dce2eeca290b3a891f1355ebfced529",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "order_item_id",
                                                "kode_produk"
                                            ],
                                            "properties": {
                                                "order_item_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "kode_produk": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "PROD-001"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kode produk diperbarui"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi kode produk gagal atau order sudah dikirim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/ship": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Input resi pengiriman",
                "operationId": "eea77259753a8f087c41903ce5f50a04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resi"
                                ],
                                "properties": {
                                    "resi": {
                                        "type": "string",
                                        "example": "JNE123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order dikirim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/run-reservation-release": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Trigger job pelepasan reservasi stok",
                "operationId": "0ec56a69f859ee7d122106714082c48d",
                "responses": {
                    "200": {
                        "description": "Job dikirim ke antrean"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Masukkan token Sanctum dengan format \"Bearer {token}\"."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}